Update comment form logic and attributes
The comment form logic has been updated to automatically set the author to the current user, and exclude the author and post from the form fields. Additionally, the bio field in the User Profile model can now be left blank, and the CKEditor5 version is updated for better compatibility.

close #16