Move UserProfile model to separate app
UserProfile model has been moved out of the blog app and into a new user_profile app. This change involved creating new files for the user_profile app which includes the model, views, templates, and tests for the UserProfile. Additionally, corresponding adjustments in settings, urls and admin registration are made to accommodate this change.