Update environment settings and hosts
This commit updates environment variables and allowed hosts in Django settings. The SECRET_KEY in the .env_sample file has been enclosed with quotes, and the ALLOWED_HOSTS list in the settings.py file has been expanded to include localhost and wildcard. The previous setting restricted the application to a single host.