Add 'unfollow' functionality to user profiles
This commit introduces the ability for users to unfollow their friends. An 'UNFOLLOW' link is now present next to each friend in the friends list on a user's profile. This also involves changes in the respective views to handle the removal of a friend from the user's profile, as well as updates to the urls for the new redirect paths after unfollowing a user. Some irrelevant and commented out code in blog/views.py were also cleaned up.

close #37