Add post detail view and display comments
A new path for the post_detail_view was added to the blog's urls. In the post_detail.html, the specific post content is now displayed along with a list of comments associated with that post. Also, a link to the post detail page was uncommented in the blog.html template.