Implement comment deletion restrictions and add friend following
This update restricts the ability to delete comments in the blog post detail template to the comment author or admins. A friends list has also been added to the blog views to aid in displaying a "Follow user" link only if the post's author is not a friend of the current user.