Add 'Read More' page functionality
A new 'Read More' HTML template has been created under the home/templates directory. The 'Read More' option in the navigation bar of the base template is updated to link to this page. The URL mapping and corresponding view function for this page have been added to 'home/urls.py' and 'home/views.py', respectively.