Update user profile page design and functionality
Significant changes have been made to the user profile page including the addition of a profile image, additional user information, and improved layout. To complement the changes, appropriate styles were also included in the CSS file. The update not only enhances the aesthetic value of the page but also adds useful features like the ability to edit the profile and unfollow other users.