Add functionality to create a new blog post
This commit provides the ability for users to create a new blog post. It includes the creation of a new form (PostForm) for creating posts, a new HTML template for the add post page, and updates in the views to handle the form submission. Accompanying URL and template updates are made as well.