Add reaction type parameter to reaction_add URL path
The reaction_add method in views.py was refactored to accept 'reaction_type' as a path parameter. This allows choosing the type of reaction (like, dislike, etc.) by clicking on the corresponding link. Code logic was added to either update the reaction type if a reaction already exists from the user or create a new reaction if it doesn't.