Add reactions functionality to blog posts
Updated the blog templates and styles to include reaction options for blog posts. A new reaction container with selectable options was added so users can react to posts with "like", "love", and other reactions. Javascript was also utilized to show or hide reaction choices upon click events. Changes were made in the views to redirect the user back to the blog home after reacting to a post.